Move _get_batch_size to utils/other.py
