+++ src/peft/peft_model.py
@@ -53,6 +53,7 @@ from .utils import (
     PeftType,
     PromptLearningConfig,
     TaskType,
+    _get_batch_size,
     _prepare_prompt_learning_config,
     _set_adapter,
     _set_trainable,
@@ -74,22 +75,6 @@ PEFT_TYPE_TO_MODEL_MAPPING = {
     PeftType.ADAPTION_PROMPT: AdaptionPromptModel,
     PeftType.IA3: IA3Model,
 }
-
-
-def _get_batch_size(input_ids: Optional[torch.Tensor], inputs_embeds: Optional[torch.Tensor]) -> int:
-    """Get the batch size based on either input_ids or input_embeds
-
-    Raises an error if both are None.
-
-    """
-    if (input_ids is None) and (inputs_embeds is None):
-        raise ValueError("You have to provide either input_ids or inputs_embeds")
-
-    if input_ids is not None:
-        batch_size = input_ids.shape[0]
-    else:
-        batch_size = inputs_embeds.shape[0]
-    return batch_size
 
 
 class PeftModel(PushToHubMixin, torch.nn.Module):
